#31: moved profile to separate plugin
