+++ at.ac.tuwien.auto.modelua.papyrus.opcua.diagram/META-INF/MANIFEST.MF
@@ -23,7 +23,8 @@ Require-Bundle: org.eclipse.papyrus.uml.extensionpoints,
  org.eclipse.gmf.runtime.draw2d.ui.render.awt;bundle-version="1.8.0",
  at.ac.tuwien.auto.modelua.papyrus.opcua.preferences;bundle-version="1.0.0",
  at.ac.tuwien.auto.modelua.papyrus.opcua.libraries;bundle-version="1.0.0",
- at.ac.tuwien.auto.modelua.papyrus.opcua.diagram.style;bundle-version="1.0.0"
+ at.ac.tuwien.auto.modelua.papyrus.opcua.diagram.style;bundle-version="1.0.0",
+ at.ac.tuwien.auto.modelua.papyrus.opcua.profile;bundle-version="1.0.0"
 Bundle-RequiredExecutionEnvironment: JavaSE-11
 Automatic-Module-Name: at.ac.tuwien.auto.modelua.papyrus.opcua.diagram
 Bundle-ActivationPolicy: lazy
