+++ at.ac.tuwien.auto.modelua.papyrus.opcua.diagram/resources/architecture/opcua.architecture
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<architecture:ArchitectureDomain xmi:version="2.0" xmlns:xmi="http://www.omg.org/XMI" xmlns:architecture="http://www.eclipse.org/papyrus/infra/core/architecture" xmlns:ecore="http://www.eclipse.org/emf/2002/Ecore" xmlns:elementtypesconfigurations="http://www.eclipse.org/papyrus/infra/elementtypesconfigurations/1.2" xmlns:gmfdiagrepresentation="http://www.eclipse.org/papyrus/infra/gmfdiag/representation" xmlns:paletteconfiguration="http://www.eclipse.org/papyrus/diagram/paletteconfiguration/0.8" xmlns:representation="http://www.eclipse.org/papyrus/infra/core/architecture/representation" xmi:id="org.eclipse.papyrus.training.systemsEngineering" id="at.ac.tuwien.auto.modelua.papyrus.opcua.systemsEngineering" name="OPC UA" description="The domain of OPC UA" icon="platform:/plugin/at.ac.tuwien.auto.modelua.papyrus.opcua.diagram/resources/images/icons/opc_ua_logo.png">
+<architecture:ArchitectureDomain xmi:version="2.0" xmlns:xmi="http://www.omg.org/XMI" xmlns:architecture="http://www.eclipse.org/papyrus/infra/core/architecture" xmlns:ecore="http://www.eclipse.org/emf/2002/Ecore" xmlns:elementtypesconfigurations="http://www.eclipse.org/papyrus/infra/elementtypesconfigurations/1.2" xmlns:gmfdiagrepresentation="http://www.eclipse.org/papyrus/infra/gmfdiag/representation" xmlns:paletteconfiguration="http://www.eclipse.org/papyrus/diagram/paletteconfiguration/0.8" xmlns:representation="http://www.eclipse.org/papyrus/infra/core/architecture/representation" xmi:id="org.eclipse.papyrus.training.systemsEngineering" id="at.ac.tuwien.auto.modelua.papyrus.opcua.systemsEngineering" name="OPC UA" description="The domain of OPC UA" icon="platform:/plugin/at.ac.tuwien.auto.modelua.papyrus.opcua.diagram.style/resources/images/icons/opc_ua_logo.png">
   <stakeholders xmi:type="architecture:Stakeholder" xmi:id="org.eclipse.papyrus.training.stackholder.librarian" id="at.ac.tuwien.auto.modelua.papyrus.opcua.stackholder.systemengineer" name="SystemEngineering" description="System Engineer of OPC UA Information Models" concerns="org.eclipse.papyrus.training.concern.requirements"/>
   <concerns xmi:type="architecture:Concern" xmi:id="org.eclipse.papyrus.training.concern.requirements" id="at.ac.tuwien.auto.modelua.papyrus.opcua.concern.systemsengineering" name="SystemEngineering" description="The concern of creating OPC UA information models"/>
-  <contexts xmi:type="architecture:ArchitectureDescriptionLanguage" xmi:id="at.ac.tuwien.auto.modelua.papyrus.opcua.architecture" id="at.ac.tuwien.auto.modelua.papyrus.opcua.architecture" name="InformationModel" description="OPC UA Informationmodel" icon="platform:/plugin/at.ac.tuwien.auto.modelua.papyrus.opcua.diagram/resources/images/icons/opc_ua_logo.png" defaultViewpoints="org.eclipse.papyrus.training.analysis" extensionPrefix="" creationCommandClass="at.ac.tuwien.auto.modelua.papyrus.opcua.diagram.architecture.OpcUaModelCreationCommand">
+  <contexts xmi:type="architecture:ArchitectureDescriptionLanguage" xmi:id="at.ac.tuwien.auto.modelua.papyrus.opcua.architecture" id="at.ac.tuwien.auto.modelua.papyrus.opcua.architecture" name="InformationModel" description="OPC UA Informationmodel" icon="platform:/plugin/at.ac.tuwien.auto.modelua.papyrus.opcua.diagram.style/resources/images/icons/opc_ua_logo.png" defaultViewpoints="org.eclipse.papyrus.training.analysis" extensionPrefix="" creationCommandClass="at.ac.tuwien.auto.modelua.papyrus.opcua.diagram.architecture.OpcUaModelCreationCommand">
     <viewpoints xmi:type="architecture:ArchitectureViewpoint" xmi:id="org.eclipse.papyrus.training.analysis" id="at.ac.tuwien.auto.modelua.papyrus.opcua.viewpoint.informationmodel" name="Information Model " description="A viewpoint allowing OPC UA information model " icon="platform:/plugin/org.eclipse.papyrus.uml.architecture/icons/viewpoint.gif" concerns="org.eclipse.papyrus.training.concern.requirements" representationKinds="org.eclipse.papyrus.training.diagram"/>
     <elementTypes xmi:type="elementtypesconfigurations:ElementTypeSetConfiguration" href="platform:/plugin/org.eclipse.papyrus.infra.emf/model/infra-emf.elementtypesconfigurations#_rWI4YHPzEeSnGJwaJWHCSg"/>
     <elementTypes xmi:type="elementtypesconfigurations:ElementTypeSetConfiguration" href="platform:/plugin/org.eclipse.papyrus.infra.gmfdiag.common/model/gmfdiag-common.elementtypesconfigurations#_rWI4YHPzEeSnGJwaJWHCSg"/>
@@ -29,7 +29,7 @@
     <elementTypes xmi:type="elementtypesconfigurations:ElementTypeSetConfiguration" href="platform:/plugin/at.ac.tuwien.auto.modelua.papyrus.opcua.diagram/resources/elementtypes/UANodeSet.di.elementtypesconfigurations#at.ac.tuwien.auto.modelua.papyrus.opcua.uaNodeSet.di.elementTypes"/>
     <elementTypes xmi:type="elementtypesconfigurations:ElementTypeSetConfiguration" href="platform:/plugin/at.ac.tuwien.auto.modelua.papyrus.opcua.diagram/resources/elementtypes/UANodeSet.elementtypesconfigurations#at.ac.tuwien.auto.modelua.papyrus.opcua.nodeset.elementTypes"/>
     <elementTypes xmi:type="elementtypesconfigurations:ElementTypeSetConfiguration" href="platform:/plugin/at.ac.tuwien.auto.modelua.papyrus.opcua.diagram/resources/elementtypes/UANodeSet.di.elementtypesconfigurations#at.ac.tuwien.auto.modelua.papyrus.opcua.nodeset.di.elementTypes"/>
-    <representationKinds xmi:type="gmfdiagrepresentation:PapyrusDiagram" xmi:id="org.eclipse.papyrus.training.diagram" id="at.ac.tuwien.auto.modelua.papyrus.opcua.diagram" name="NodeSetDiagram" icon="platform:/plugin/at.ac.tuwien.auto.modelua.papyrus.opcua.diagram/resources/images/icons/opc_ua_logo.png" implementationID="PapyrusUMLClassDiagram" customStyle="platform:/plugin/at.ac.tuwien.auto.modelua.papyrus.opcua.diagram.style/resources/css/UANodeSet.css" creationCommandClass="org.eclipse.papyrus.uml.diagram.clazz.CreateClassDiagramCommand">
+    <representationKinds xmi:type="gmfdiagrepresentation:PapyrusDiagram" xmi:id="org.eclipse.papyrus.training.diagram" id="at.ac.tuwien.auto.modelua.papyrus.opcua.diagram" name="NodeSetDiagram" icon="platform:/plugin/at.ac.tuwien.auto.modelua.papyrus.opcua.diagram.style/resources/images/icons/opc_ua_logo.png" implementationID="PapyrusUMLClassDiagram" customStyle="platform:/plugin/at.ac.tuwien.auto.modelua.papyrus.opcua.diagram.style/resources/css/UANodeSet.css" creationCommandClass="org.eclipse.papyrus.uml.diagram.clazz.CreateClassDiagramCommand">
       <concerns xmi:type="architecture:Concern" href="platform:/plugin/org.eclipse.papyrus.uml.architecture/model/uml.architecture#_HQhf5MSYEeaKZJ_pGfaSiA"/>
       <modelRules xmi:type="representation:ModelRule" xmi:id="_ZrBf4ZaBEeS8eNvSwD0lgA" permit="true" elementMultiplicity="1" multiplicity="-1">
         <element xmi:type="ecore:EClass" href="http://www.eclipse.org/uml2/5.0.0/UML#//Package"/>
