--- at.ac.tuwien.auto.modelua.papyrus.opcua.diagram/resources/css/UANodeSet.css
@@ -1,174 +1,0 @@
-
-Class {
-	elementIcon: false;
-	displayBorder: false;
-	transparency:100;
-	showTitle:false;
-	displayHeader:false;
-	followSVGSymbol:true;
-	maintainSymbolRatio:true;
-}
-
-Class > Compartment{
-	visible:false;
-}
- 
-Class > Compartment[kind=symbol]{
-	visible: true;
-}
-
-Label[type=StereotypeLabel] {
-	visible:false;
-}
-
-[appliedStereotypes~="NodeSet::UAObject"] {
-	italic:false;
-	svgFile: "platform:/plugin/at.ac.tuwien.auto.modelua.papyrus.opcua.diagram/resources/images/shapes/UAObject.svg";
-}
-
-[appliedStereotypes~="NodeSet::UAObjectType"] {
-	svgFile: "platform:/plugin/at.ac.tuwien.auto.modelua.papyrus.opcua.diagram/resources/images/shapes/UAObjectType.svg";
-}
-
-[appliedStereotypes~="NodeSet::UAVariable"] {
-	svgFile: "platform:/plugin/at.ac.tuwien.auto.modelua.papyrus.opcua.diagram/resources/images/shapes/UAVariable.svg";
-}
-
-[appliedStereotypes~="NodeSet::UAVariableType"] {
-	italic:true;
-	svgFile: "platform:/plugin/at.ac.tuwien.auto.modelua.papyrus.opcua.diagram/resources/images/shapes/UAVariableType.svg";
-}
-
-[appliedStereotypes~="NodeSet::UADataType"] {
-	svgFile: "platform:/plugin/at.ac.tuwien.auto.modelua.papyrus.opcua.diagram/resources/images/shapes/UADataType.svg";
-}
-
-[appliedStereotypes~="NodeSet::UAReferenceType"]{
-	svgFile: "platform:/plugin/at.ac.tuwien.auto.modelua.papyrus.opcua.diagram/resources/images/shapes/UAReferenceType.svg";
-}
-
-[appliedStereotypes~="NodeSet::UAMethod"] {
-	svgFile: "platform:/plugin/at.ac.tuwien.auto.modelua.papyrus.opcua.diagram/resources/images/shapes/UAMethod.svg";
-}
-
-[appliedStereotypes~="NodeSet::UAView"] {
-	svgFile: "platform:/plugin/at.ac.tuwien.auto.modelua.papyrus.opcua.diagram/resources/images/shapes/UAView.svg";
-}
-
-/*StereotypeComment{
-   visible: true;
-}
-*/
-
-StereotypeComment Compartment[type=StereotypeBrace]{
-	visible:false;
-}
-
-StereotypeComment Compartment[type=StereotypeCompartment][stereotype="NodeSet::Reference"]{
-   visible: true;
-}
-
-StereotypeComment Compartment[type=StereotypeCompartment][stereotype="NodeSet::Reference"]>[property="referenceType"]{
-   visible: true;
-}
-
-StereotypeComment Compartment[type=StereotypeCompartment][stereotype="NodeSet::Reference"]>[property="isForward"]{
-   visible: false;
-}
-
-StereotypeComment Compartment[type=StereotypeCompartment][stereotype="NodeSet::Reference"]>[property="referenceType_symmetric"]{
-   visible: false;
-}
-
-StereotypeComment Compartment[type=StereotypeCompartment][stereotype="NodeSet::Reference"]>[property="referenceType_browseName"]{
-   visible: false;
-}
-
-StereotypeComment Compartment[type=StereotypeCompartment][stereotype="NodeSet::Reference"]>[property="referenceType_isHierachical"]{
-   visible: false;
-}
-
-
-
-/* General Decorators for References */
-[appliedStereotypes~="NodeSet::Reference"][referenceType_symmetric=true] {
-	sourceDecoration: "ReferenceTypeDecorator";
-	targetDecoration: "ReferenceTypeDecorator";
-}
-
-[appliedStereotypes~="NodeSet::Reference"][referenceType_symmetric=false][referenceType_isHierachical=false][isForward=false] {
-	sourceDecoration:"ReferenceTypeDecorator";	
-	targetDecoration:"none";
-}
-
-[appliedStereotypes~="NodeSet::Reference"][referenceType_symmetric=false][referenceType_isHierachical=false][isForward=true] {
-	sourceDecoration:"none";
-	targetDecoration:"ReferenceTypeDecorator";
-}
-
-/* Decorators for Hierarchical References */
-[appliedStereotypes~="NodeSet::Reference"][referenceType_symmetric=false][referenceType_isHierachical=true][isForward=false] {
-	sourceDecoration:"HierachicalReferenceTypeDecorator";
-	targetDecoration:"none";
-}
-
-[appliedStereotypes~="NodeSet::Reference"][referenceType_symmetric=false][referenceType_isHierachical=true][isForward=true] {
-	sourceDecoration:"none";
-	targetDecoration:"HierachicalReferenceTypeDecorator";
-}
-
-/* Decorators for HasComponent References */
-[appliedStereotypes~="NodeSet::Reference"][referenceType_symmetric|=""][referenceType_browseName~="HasComponent"][isForward=false] {
-	sourceDecoration:"HasComponentDecorator";	
-	targetDecoration:"none";
-}
-
-[appliedStereotypes~="NodeSet::Reference"][referenceType_symmetric|=""][referenceType_browseName~="HasComponent"][isForward=true] {
-	sourceDecoration:"none";
-	targetDecoration: "HasComponentDecorator";
-}
-
-/* Decorators for HasProperty References */
-[appliedStereotypes~="NodeSet::Reference"][referenceType_symmetric|=""][referenceType_browseName~="HasProperty"][isForward=false] {
-	sourceDecoration:"HasPropertyDecorator";	
-	targetDecoration:"none";
-}
-
-[appliedStereotypes~="NodeSet::Reference"][referenceType_symmetric|=""][referenceType_browseName~="HasProperty"][isForward=true] {
-	sourceDecoration:"none";
-	targetDecoration:"HasPropertyDecorator";
-}
-
-/* Decorators for HasTypeDefinition References */
-[appliedStereotypes~="NodeSet::Reference"][referenceType_symmetric|=""][referenceType_browseName~="HasTypeDefinition"][isForward=false] {
-	sourceDecoration:"HasTypeDefinitionDecorator";	
-	targetDecoration:"none";
-}
-
-[appliedStereotypes~="NodeSet::Reference"][referenceType_symmetric|=""][referenceType_browseName~="HasTypeDefinition"][isForward=true] {
-	sourceDecoration:"none";
-	targetDecoration: "HasTypeDefinitionDecorator";
-}
-
-/* Decorators for HasSubtype References */
-/* Attention HasSubtype is inverted!! */
-[appliedStereotypes~="NodeSet::Reference"][referenceType_symmetric|=""][referenceType_browseName~="HasSubtype"][isForward=false] {
-	sourceDecoration:"none";
-	targetDecoration: "HasSubtypeDecorator";
-}
-
-[appliedStereotypes~="NodeSet::Reference"][referenceType_symmetric|=""][referenceType_browseName~="HasSubtype"][isForward=true] {
-	sourceDecoration:"HasSubtypeDecorator";	
-	targetDecoration:"none";
-}
-
-/* Decorators for HasEventSource References */
-[appliedStereotypes~="NodeSet::Reference"][referenceType_symmetric|=""][referenceType_browseName~="HasEventSource"][isForward=false] {
-	sourceDecoration:"HasEventSourceDecorator";	
-	targetDecoration:"none";
-}
-
-[appliedStereotypes~="NodeSet::Reference"][referenceType_symmetric|=""][referenceType_browseName~="HasEventSource"][isForward=true] {
-	sourceDecoration:"none";
-	targetDecoration: "HasEventSourceDecorator";
-}
