+++ at.ac.tuwien.auto.modelua.papyrus.opcua.profile/.gitignore
@@ -1,0 +1,1 @@
+/bin/
