+++ at.ac.tuwien.auto.modelua.papyrus.opcua.profile/META-INF/MANIFEST.MF
@@ -1,0 +1,12 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: at.ac.tuwien.auto.modelua.papyrus.opcua.profile
+Bundle-SymbolicName: at.ac.tuwien.auto.modelua.papyrus.opcua.profile;singleton:=true
+Bundle-Version: 1.0.0.qualifier
+Bundle-Activator: at.ac.tuwien.auto.modelua.papyrus.opcua.profile.Activator
+Require-Bundle: org.eclipse.ui,
+ org.eclipse.core.runtime
+Bundle-RequiredExecutionEnvironment: JavaSE-11
+Automatic-Module-Name: at.ac.tuwien.auto.modelua.papyrus.opcua.profile
+Bundle-ActivationPolicy: lazy
+Export-Package: at.ac.tuwien.auto.modelua.papyrus.opcua.profile
