+++ at.ac.tuwien.auto.modelua.papyrus.opcua.profile/build.properties
@@ -1,0 +1,5 @@
+source.. = src/
+output.. = bin/
+bin.includes = META-INF/,\
+               .,\
+               plugin.xml
